feat(inflation): refactor keeper methods into several files
