+++ x/inflation/keeper/minter.go
@@ -1,0 +1,57 @@
+package keeper
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+
+	"github.com/tharsis/evmos/x/inflation/types"
+)
+
+// get the minter
+func (k Keeper) GetMinter(ctx sdk.Context) (minter types.Minter) {
+	store := ctx.KVStore(k.storeKey)
+	b := store.Get(types.MinterKey)
+	if b == nil {
+		panic("stored minter should not have been nil")
+	}
+
+	k.cdc.MustUnmarshal(b, &minter)
+	return
+}
+
+// set the minter
+func (k Keeper) SetMinter(ctx sdk.Context, minter types.Minter) {
+	store := ctx.KVStore(k.storeKey)
+	b := k.cdc.MustMarshal(&minter)
+	store.Set(types.MinterKey, b)
+}
+
+// GetLastHalvenEpochNum returns last halven epoch number
+func (k Keeper) GetLastHalvenEpochNum(ctx sdk.Context) int64 {
+	store := ctx.KVStore(k.storeKey)
+	b := store.Get(types.LastHalvenEpochKey)
+	if b == nil {
+		return 0
+	}
+
+	return int64(sdk.BigEndianToUint64(b))
+}
+
+// SetLastHalvenEpochNum set last halven epoch number
+func (k Keeper) SetLastHalvenEpochNum(ctx sdk.Context, epochNum int64) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.LastHalvenEpochKey, sdk.Uint64ToBigEndian(uint64(epochNum)))
+}
+
+// CreateDeveloperVestingModuleAccount creates the module account for developer vesting.
+func (k Keeper) CreateDeveloperVestingModuleAccount(ctx sdk.Context, amount sdk.Coin) {
+	moduleAcc := authtypes.NewEmptyModuleAccount(
+		types.DeveloperVestingModuleAcctName, authtypes.Minter)
+
+	k.accountKeeper.SetModuleAccount(ctx, moduleAcc)
+
+	err := k.bankKeeper.MintCoins(ctx, types.DeveloperVestingModuleAcctName, sdk.NewCoins(amount))
+	if err != nil {
+		panic(err)
+	}
+}
